requirements.txt
Django
psycopg2
python-dotenv
channels
daphne
djangorestframework